implement AllTeams component
