+++ client/src/pages/TeamsPage.js
@@ -1,17 +1,13 @@
 import { NavLink } from "react-router-dom";
 import { useLoaderData } from "react-router-dom";
-import { useState } from "react";
 import axios from "axios";
-import sortTeams from "../utils/sortTeams";
 import ScrollableList from "../components/ScrollableList";
 import AuthedPageTitle from "../components/AuthedPageTitle";
 import CreateTeamButton from "../components/CreateTeamButton";
+import AllTeams from "../components/AllTeams";
 
 export const TeamsPage = () => {
-  const { teams, userTeams } = useLoaderData();
-
-  const [sortBy, setSortBy] = useState("none");
-  const sortedTeams = sortTeams(teams, sortBy);
+  const { userTeams } = useLoaderData();
 
   return (
     <>
@@ -21,28 +17,7 @@ export const TeamsPage = () => {
       </div>
       <div className="flex sm:flex-row flex-col w-full gap-10 h-full mt-8 pb-2 overflow-hidden">
         <div className="sm:w-3/4 sm:h-full h-60">
-          <ScrollableList
-            title="Teams"
-            hasSortBy={true}
-            sortBy={sortBy}
-            setSortBy={setSortBy}
-          >
-            {sortedTeams.map((team, index) => (
-              <NavLink
-                to={`/teams/${team.id}`}
-                className="flex justify-between text-xs sm:text-base bg-white p-2.5 border-t-[0.5px] border-l-[0.5px] rounded-sm shadow-[0_0.3px_1px_rgba(0,0,0,0.2)] hover:bg-blue-200"
-                key={`${team.name}-${index}`}
-              >
-                <div className="flex gap-1">
-                  <p className="font-semibold">{team.name} /</p>
-                  <p>{team.jobField}</p>
-                </div>
-                <div className="flex w-6 h-6 rounded-full items-center justify-center text-white bg-blue-800 text-xs">
-                  {team.userCount}
-                </div>
-              </NavLink>
-            ))}
-          </ScrollableList>
+          <AllTeams />
         </div>
         <div className="sm:h-full sm:w-1/4 h-60">
           <ScrollableList title="Your Teams">
